+++ docs/tutorials/tabular_prediction/tabular-kaggle.md
@@ -39,7 +39,7 @@ This means you'll need to run the above steps with `[COMPETITION]` replaced by `
 ```
 import pandas as pd
 import numpy as np
-from autogluon.tabular import TabularPredictorV2
+from autogluon.tabular import TabularPredictor
 
 directory = '~/IEEEfraud/'  # directory where you have downloaded the data CSV files from the competition
 label_column = 'isFraud'  # name of target variable to predict in this competition
@@ -58,7 +58,7 @@ train_data = pd.merge(train_transaction, train_identity, on='TransactionID', how
 
 Note that a left-join on the `TransactionID` key happened to be most appropriate for this Kaggle competition, but for others involving multiple training data files, you will likely need to use a different join strategy (always consider this very carefully). Now that all our training data resides within a single table, we can apply AutoGluon. Below, we specify the `presets` argument to maximize AutoGluon's predictive accuracy which usually requires that you run `fit()` with longer time limits (3600s below should likely be increased in your run):
 ```
-predictor = TabularPredictorV2(label=label_column, eval_metric=eval_metric, path=save_path, verbosity=3).fit(
+predictor = TabularPredictor(label=label_column, eval_metric=eval_metric, path=save_path, verbosity=3).fit(
     train_data, presets='best_quality', time_limit=3600
 )
